Add LockedUntil column
If a job's connection dies, then the advisory lock will be released and
another worker can lock the job and start working it, leading to a
situation where multiple workers could be working the same job.

This change minimises that risk by adding a locked_until column which is
bumped in a goroutine whilst a job is being worked, so even if the
advisory lock goes away, the job can remain locked until it is done.

Signed-off-by: Lewis Marshall <lewis@lmars.net>